Placeholder navbar and pages
A "normal" top navbar for navigation through pages, missing a logo for navigating to home page. Also, added empty pages for every link on the navbar. Some styling adjustments, choosing fonts etc. Made a `Layout` component which surrounds every page with a navbar and in the future with a footer. The website is for now kinda-responsive.