Blog pages, finally!
Blog pages now work when clicking on `Read more`, yay! It was quite a hassle to setup rendering of `.mdx` files but it now works. There's a lot of work to do with styling the text as well as maybe implement some kind of code highlighting. Still a lot of CSS work to do on that part. Seems like React 18 doesn't work well with next-mdx-remote, as I can't use react elements in the markdown, I should research about it more.